Re-ordered FAQ in 'Building with CMake' page.

+++ ncdump/nciter.c
@@ -288,7 +288,7 @@ nc_next_iter(nciter_t *iter,	/* returned opaque iteration state */
 					    iter->chunksizes, start);
 	    /* adjust count to stay in range of dimsizes */
 	    for(i = 0; i < iter->rank; i++) {
-		int leftover = iter->dimsizes[i] - start[i];
+		size_t leftover = iter->dimsizes[i] - start[i];
 		if(iter->chunksizes[i] <= iter->dimsizes[i])
 		    count[i] = iter->chunksizes[i];
 		else /* can happen for variables with only unlimited dimensions */
